Models, Routes, Migrations, & Factories Created
In this update, models modified and made eloquent. Products has been made associated with categories and users.
Migrations added to code. Needed attributes are added as well.
For ease of use and automising the process, factories are added to generate some data.
Can not provide the view due to the error 'Undefined variable $products'.